requirements.txt
mlx-whisper>=0.4.3
openai>=1.0.0
python-dotenv>=1.0.1
